fix: code refactoring, optimisation
